Day5/Quiz Assistant/requirements.txt
PyPDF2
langchain
google-generativeai
streamlit